feat: add code coverage integration and windows build ci
At first I tried to use OpenCover to calculate code coverage. But it doesn't run on Linux and faced some trouble running the script on Windows machine too. So now using coverlet for code coverage. So far working good. 

Although Web project's coverage is not calculated. May be because it is not referenced by the UnitTests test project 

Integrate with AppVeyor to run the build on Windows OS

Also update readme file to show new nice cool badges. 

closes #21